# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /registerNewCustomer:
    post:
      tags:
      - customer-controller
      operationId: registerNewCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
...
  /users/{id}:
    get:
      tags:
      - customer-controller
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    CurrentAccount:
      type: object
      properties:
        id:
          type: integer
          format: int32
        balance:
          type: number
          format: double
        account_number:
          type: integer
          format: int32
    SavingsAccount:
      type: object
      properties:
        id:
          type: integer
          format: int32
        balance:
          type: number
          format: double
        account_number:
          type: integer
          format: int32
    User:
      type: object
      properties:
        user_id:
          type: integer
          format: int32
        aadhar_id:
          type: string
        enabled:
          type: string
          format: byte
        first_Name:
          type: string
        last_Name:
          type: string
        password:
          type: string
        phone:
          type: string
        username:
          type: string
        savingsAccount:
          $ref: '#/components/schemas/SavingsAccount'
        currentAccount:
          $ref: '#/components/schemas/CurrentAccount'
        savingsAccountId:
          type: integer
          format: int32
        currentAccountId:
          type: integer
          format: int32